Ensure crop points are current
Previously setting the crop points via the text entry dialog would
update the sliders, but moving the sliders would not update the
pre-populated values in the text entry dialog. This was due to the text
field being populated with the `Trial` values, which aren't updated
until the crop changes are saved. This change uses the crop values from
the current playback view controller, which is the same value that the
sliders present and modify.

Fixes #47