Introduce Action Area state
The Action Area can present detail screens that have either a persistent set of
actions or two sets that are toggled between when a primary action is taken.
This change introduces enums for these action configurations and a
corresponding state that will be used to control detail dismissal behavior in a
future change.

PiperOrigin-RevId: 256974808