Fix flaky test related to SensorDataManager
One of the `SensorDataManager` methods accepted an optional completion block,
which made it easy for clients to unintentionally not wait for that method to
complete its async behavior. This was most likely causing a race condition
between async behavior initiated during tests and test teardown. This change
turns such an error into a noop, which should only happen during test
execution, and if it were to happen in the real system, it would be redundant
because the store would already have been truncated.

PiperOrigin-RevId: 259431059