Improve Action Area API
These changes add what should be the last missing functionality for v1. The
main API for presenting content view controllers is now also the same as
`UINavigationController` and `UISplitViewController`, which will allow it to be
used interchangeably with simpler content view controllers that don't have any
Action Area specific behavior.

* Add support for detail empty state view controllers
* Adopt the standard `UIKit` container API
* Move Action Area specific details to protocols
* Provide standard containers for master / detail content
* Move all core types under the same namespace.

PiperOrigin-RevId: 262178443